* add 'show_comments' field to interfaces  * integrate show_comments to portlet initialization  * add protected display of comment count to individual entries
The comment count is dependent on the presence of the <slash:comments/> element in the source feed.  If this element is present, and the setting is True (default is False), then comment counts will be displayed for each listed feed entry.  If the setting is True, but the element is __not__ present, then the comment count will be omitted.